refactor(code_gen): extract string builtins into builtins/string.rs
