+++ src/code_gen/builtins/mod.rs
@@ -2,6 +2,7 @@
 
 mod core;
 mod io;
+mod string;
 
 use super::{Interpreter, Value};
 use crate::ast::Expression;
@@ -41,6 +42,9 @@ impl Interpreter {
             return Ok(Some(v));
         }
         if let Some(v) = io::call(self, name, args)? {
+            return Ok(Some(v));
+        }
+        if let Some(v) = string::call(self, name, args)? {
             return Ok(Some(v));
         }
         match name {
@@ -287,42 +291,6 @@ impl Interpreter {
 
                 let now = Local::now();
                 Ok(Some(Value::String(now.format("%Y-%m-%d").to_string())))
-            }
-
-            "parse_int" => {
-                if args.len() != 1 {
-                    return Err(format!(
-                        "parse_int() expects 1 argument, got {}",
-                        args.len()
-                    ));
-                }
-
-                let val = self.eval_expression(&args[0])?;
-                if let Value::String(s) = val {
-                    match s.parse::<i64>() {
-                        Ok(n) => Ok(Some(Value::Int(n))),
-                        Err(_) => Ok(Some(Value::Int(0))),
-                    }
-                } else {
-                    Err("parse_int() expects a string argument".to_string())
-                }
-            }
-
-            "char_code" => {
-                if args.len() != 1 {
-                    return Err(format!(
-                        "char_code() expects 1 argument, got {}",
-                        args.len()
-                    ));
-                }
-
-                let val = self.eval_expression(&args[0])?;
-                if let Value::String(s) = val {
-                    let code = s.chars().next().map(|c| c as i64).unwrap_or(0);
-                    Ok(Some(Value::Int(code)))
-                } else {
-                    Err("char_code() expects a string argument".to_string())
-                }
             }
 
             "sys_timestamp" => {
